+++ packaging/rpm/build-rpm.sh
@@ -156,6 +156,7 @@ fi
 
 install -d -o "\$service_user" -g "\$service_group" -m 0750 /run/dasobjectstore
 install -d -o "\$service_user" -g "\$service_group" -m 0750 /var/lib/dasobjectstore
+install -d -o "\$service_user" -g "\$service_group" -m 0750 /var/lib/dasobjectstore/report-rebuild
 install -d -o "\$service_user" -g "\$service_group" -m 0750 /var/log/dasobjectstore
 install -d -o "\$service_user" -g "\$service_group" -m 0750 /opt/dasobjectstore
 install -d -o "\$service_user" -g "\$service_group" -m 0750 /opt/dasobjectstore/tls